requirements.txt
Flask==2.1.2
flask_sqlalchemy==2.5.1
psycopg2-binary==2.9.3
python-dateutil==2.8.2
flask_swagger_ui==4.11.1
urllib3==1.26.9
python-dotenv==0.20.0
gunicorn==20.1.0